--- WebPlaylistToSpotify/SystemBrowser.cs
@@ -1,209 +1,0 @@
-﻿using Microsoft.AspNetCore.Builder;
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.AspNetCore.Http;
-using SpotifyAPI.Web;
-using System.Diagnostics;
-using System.Net;
-using System.Net.Sockets;
-using System.Runtime.InteropServices;
-using System.Text;
-
-namespace WebPlaylistToSpotify
-{
-    public class SystemBrowser
-    {
-        public int Port { get; }
-        private readonly string _path;
-        private string _verifier;
-        private string _usedCallbackUri;
-
-        public SystemBrowser(int? port = null, string path = null)
-        {
-            _path = path;
-
-            if (!port.HasValue)
-            {
-                Port = GetRandomUnusedPort();
-            }
-            else
-            {
-                Port = port.Value;
-            }
-        }
-
-        private int GetRandomUnusedPort()
-        {
-            var listener = new TcpListener(IPAddress.Loopback, 0);
-            listener.Start();
-            var port = ((IPEndPoint)listener.LocalEndpoint).Port;
-            listener.Stop();
-            return port;
-        }
-
-        private string AuthUrl(string callbackUrl)
-        {
-            var (verifier, challenge) = PKCEUtil.GenerateCodes();
-            _verifier = verifier;
-
-            var loginRequest = new SpotifyAPI.Web.LoginRequest(
-              new Uri(callbackUrl),
-              "cecbc33419334a7e968eddbd26639cc0",
-              SpotifyAPI.Web.LoginRequest.ResponseType.Code
-            )
-            {
-                CodeChallengeMethod = "S256",
-                CodeChallenge = challenge,
-                Scope = new[] { Scopes.PlaylistModifyPublic }
-            };
-            return loginRequest.ToUri().ToString();
-        }
-
-        public async Task<string> GetToken(string code)
-        {
-            var initialResponse = await new OAuthClient().RequestToken(
-                new PKCETokenRequest("cecbc33419334a7e968eddbd26639cc0", code, new Uri(_usedCallbackUri), _verifier)
-            );
-
-            return initialResponse.AccessToken;
-        }
-
-        public async Task<string> InvokeAsync(CancellationToken cancellationToken = default(CancellationToken))
-        {
-            using (var listener = new LoopbackHttpListener(Port, _path))
-            {
-                _usedCallbackUri = listener.Url;
-                var authUrl = AuthUrl(listener.Url);
-                OpenBrowser(authUrl);
-
-                return await listener.WaitForCallbackAsync();
-            }
-        }
-
-        public static void OpenBrowser(string url)
-        {
-            try
-            {
-                Process.Start(url);
-            }
-            catch
-            {
-                // hack because of this: https://github.com/dotnet/corefx/issues/10361
-                if (RuntimeInformation.IsOSPlatform(OSPlatform.Windows))
-                {
-                    url = url.Replace("&", "^&");
-                    Process.Start(new ProcessStartInfo("cmd", $"/c start {url}") { CreateNoWindow = true });
-                }
-                else if (RuntimeInformation.IsOSPlatform(OSPlatform.Linux))
-                {
-                    Process.Start("xdg-open", url);
-                }
-                else if (RuntimeInformation.IsOSPlatform(OSPlatform.OSX))
-                {
-                    Process.Start("open", url);
-                }
-                else
-                {
-                    throw;
-                }
-            }
-        }
-    }
-
-    public class LoopbackHttpListener : IDisposable
-    {
-        const int DefaultTimeout = 60 * 5; // 5 mins (in seconds)
-
-        IWebHost _host;
-        TaskCompletionSource<string> _source = new TaskCompletionSource<string>();
-        string _url;
-
-        public string Url => _url;
-
-        public LoopbackHttpListener(int port, string path = null)
-        {
-            path = path ?? String.Empty;
-            if (path.StartsWith("/")) path = path.Substring(1);
-
-            _url = $"http://127.0.0.1:{port}/{path}";
-
-            _host = new WebHostBuilder()
-                .UseKestrel()
-                .UseUrls(_url)
-                .Configure(Configure)
-                .Build();
-            _host.Start();
-        }
-
-        public void Dispose()
-        {
-            Task.Run(async () =>
-            {
-                await Task.Delay(500);
-                _host.Dispose();
-            });
-        }
-
-        void Configure(IApplicationBuilder app)
-        {
-            app.Run(async ctx =>
-            {
-                if (ctx.Request.Method == "GET")
-                {
-                    await SetResultAsync(ctx.Request.QueryString.Value, ctx);
-                }
-                else if (ctx.Request.Method == "POST")
-                {
-                    if (!ctx.Request.ContentType.Equals("application/x-www-form-urlencoded", StringComparison.OrdinalIgnoreCase))
-                    {
-                        ctx.Response.StatusCode = 415;
-                    }
-                    else
-                    {
-                        using (var sr = new StreamReader(ctx.Request.Body, Encoding.UTF8))
-                        {
-                            var body = await sr.ReadToEndAsync();
-                            await SetResultAsync(body, ctx);
-                        }
-                    }
-                }
-                else
-                {
-                    ctx.Response.StatusCode = 405;
-                }
-            });
-        }
-
-        private async Task SetResultAsync(string value, HttpContext ctx)
-        {
-            try
-            {
-                ctx.Response.StatusCode = 200;
-                ctx.Response.ContentType = "text/html";
-                await ctx.Response.WriteAsync("<h1>You can now return to the application.</h1>");
-                await ctx.Response.Body.FlushAsync();
-
-                _source.TrySetResult(value);
-            }
-            catch (Exception ex)
-            {
-                Console.WriteLine(ex.ToString());
-
-                ctx.Response.StatusCode = 400;
-                ctx.Response.ContentType = "text/html";
-                await ctx.Response.WriteAsync("<h1>Invalid request.</h1>");
-                await ctx.Response.Body.FlushAsync();
-            }
-        }
-
-        public Task<string> WaitForCallbackAsync(int timeoutInSeconds = DefaultTimeout)
-        {
-            Task.Run(async () =>
-            {
-                await Task.Delay(timeoutInSeconds * 1000);
-                _source.TrySetCanceled();
-            });
-
-            return _source.Task;
-        }
-    }
-}
